+++ legacy_packages/chains/chains/61.ts
@@ -53,13 +53,14 @@ export default {
     "symbol": "ETC",
     "decimals": 18
   },
-  "networkId": 1,
+  "networkId": 61,
+  "redFlags": [],
   "rpc": [
     "https://61.rpc.thirdweb.com/${THIRDWEB_API_KEY}",
+    "https://etc.etcdesktop.com",
     "https://etc.rivet.link",
     "https://besu-at.etc-network.info",
     "https://geth-at.etc-network.info",
-    "https://etc.etcdesktop.com",
     "https://etc.mytokenpocket.vip"
   ],
   "shortName": "etc",