+++ legacy_packages/react/src/components/OTPInput.tsx
@@ -1,140 +1,109 @@
-import { useEffect, useRef } from "react";
-import { Input } from "./formElements";
-import { Container } from "./basic";
-import { media, fontSize, spacing } from "../design-system";
-import styled from "@emotion/styled";
+"use client";
+import { keyframes } from "@emotion/react";
+import { OTPInput as InputOTP, type SlotProps } from "input-otp";
+import { fontSize, spacing, radius } from "../design-system";
 import { useCustomTheme } from "../design-system/CustomThemeProvider";
+import { StyledDiv } from "../design-system/elements";
+import { Container } from "./basic";
 
+/**
+ * @internal
+ */
 export function OTPInput(props: {
   digits: number;
   isInvalid?: boolean;
   value: string;
   setValue: (value: string) => void;
-  onEnter?: () => void;
+  onEnter: () => void;
 }) {
-  const otp = props.value.split("").map(Number);
-
-  const setOTP = (newOTP: number[]) => {
-    props.setValue(newOTP.join(""));
-  };
-
-  const inputToFocusIndex = otp.length;
-  const boxEls = useRef<(HTMLInputElement | null)[]>([]);
-
-  useEffect(() => {
-    if (boxEls.current[inputToFocusIndex]) {
-      requestAnimationFrame(() => {
-        boxEls.current[inputToFocusIndex]?.focus();
-      });
-    }
-  }, [inputToFocusIndex]);
-
   return (
-    <Container center="x" gap="sm" flex="row">
-      {new Array(props.digits).fill(null).map((_, i) => {
-        return (
-          <OTPInputBox
-            data-error={props.isInvalid}
-            ref={(e) => (boxEls.current[i] = e)}
-            key={i}
-            value={otp[i] ?? ""}
-            type="number"
-            pattern="[0-9]*"
-            variant="outline"
-            inputMode="numeric"
-            onPaste={(e) => {
-              const pastedData = e.clipboardData.getData("text/plain");
-              const newOTP = pastedData
-                .slice(0, props.digits)
-                .split("")
-                .filter((n) => /\d/.test(n))
-                .map(Number);
-
-              setOTP(newOTP);
-              e.preventDefault();
-            }}
-            onKeyDown={(e) => {
-              if (e.key === "Enter") {
-                if (props.onEnter) {
-                  props.onEnter();
-                  return;
-                }
-              }
-
-              if (e.key === "ArrowLeft") {
-                if (i === 0) {
-                  return;
-                }
-
-                boxEls.current[i - 1]?.focus();
-                return;
-              }
-
-              if (e.key === "ArrowRight") {
-                if (i === props.digits - 1) {
-                  return;
-                }
-
-                boxEls.current[i + 1]?.focus();
-                return;
-              }
-
-              if (e.key === "e" || e.key === ".") {
-                e.preventDefault();
-                return;
-              }
-
-              if (e.key === "Backspace") {
-                if (i === 0) {
-                  return;
-                }
-
-                const newOTP = otp.slice(0, -1);
-                setOTP(newOTP);
-              }
-            }}
-            onChange={(e) => {
-              const value = e.target.value;
-
-              if (value.length > 1) {
-                setOTP(value.split("").filter((n) => /\d/.test(n)).map(Number));
-                return;
-              }
-
-              if (!/\d/.test(value) && value !== "") {
-                e.preventDefault();
-                return;
-              }
-
-              const newOTP = [...otp];
-              const index = i > inputToFocusIndex - 1 ? inputToFocusIndex : i;
-
-              newOTP[index] = Number(value);
-              setOTP(newOTP);
-            }}
-          />
-        );
-      })}
-    </Container>
+    <OTPInputContainer>
+      <InputOTP
+        onComplete={() => {
+          props.onEnter();
+        }}
+        maxLength={6}
+        value={props.value}
+        render={({ slots }) => (
+          <Container flex="row" gap="xs" center="both">
+            {slots.map((slot, idx) => (
+              <Slot key={idx} {...slot} isInvalid={props.isInvalid} />
+            ))}
+          </Container>
+        )}
+        onChange={(newValue) => {
+          props.setValue(newValue);
+        }}
+      />
+    </OTPInputContainer>
   );
 }
 
-const OTPInputBox = /* @__PURE__ */ styled(Input)(() => {
+const OTPInputContainer = /* @__PURE__ */ StyledDiv({
+  "& input": {
+    maxWidth: "100%",
+  },
+});
+
+function Slot(props: SlotProps & { isInvalid?: boolean }) {
+  return (
+    <OTPInputBox data-active={props.isActive} data-error={props.isInvalid}>
+      {props.char !== null && <div>{props.char}</div>}
+      {props.hasFakeCaret && <FakeCaret />}
+    </OTPInputBox>
+  );
+}
+
+const caretBlink = keyframes`
+  0%, 100% {
+    opacity: 0;
+  },
+  50% {
+    opacity: 1;
+  }
+  `;
+
+const FakeCaret = /* @__PURE__ */StyledDiv(() => {
   const theme = useCustomTheme();
   return {
-    appearance: "none",
-    WebkitAppearance: "none",
+    position: "absolute",
+    pointerEvents: "none",
+    inset: 0,
+    display: "flex",
+    alignItems: "center",
+    justifyContent: "center",
+    animation: `${caretBlink} 1s infinite`,
+    "&::before": {
+      content: "''",
+      display: "block",
+      width: "1.5px",
+      height: "1em",
+      backgroundColor: theme.colors.primaryText,
+    },
+  };
+});
+
+const OTPInputBox = /* @__PURE__ */ StyledDiv(() => {
+  const theme = useCustomTheme();
+  return {
+    position: "relative",
     width: "40px",
     height: "40px",
     textAlign: "center",
+    display: "flex",
+    alignItems: "center",
+    justifyContent: "center",
     fontSize: fontSize.md,
     padding: spacing.xs,
-    [media.mobile]: {
-      width: "35px",
-      height: "35px",
+    boxSizing: "border-box",
+    transition: "color 200ms ease",
+    border: `2px solid ${theme.colors.borderColor}`,
+    "&[data-active='true']": {
+      borderColor: theme.colors.accentText,
     },
-    "&[data-verify-status='invalid']": {
-      color: theme.colors.danger,
+    color: theme.colors.primaryText,
+    borderRadius: radius.lg,
+    "&[data-error='true']": {
       borderColor: theme.colors.danger,
     },
   };