+++ packages/thirdweb/src/utils/ipfs.ts
@@ -34,11 +34,18 @@ export function resolveScheme(options: ResolveSchemeOptions) {
     const clientId = options.client.clientId;
     const cid = findIPFSCidFromUri(options.uri);
 
+    let bundleId: string | undefined = undefined;
+    if (typeof globalThis !== "undefined" && "Application" in globalThis) {
+      // shims use wallet connect RN module which injects Application info in globalThis
+      // biome-ignore lint/suspicious/noExplicitAny: get around globalThis typing
+      bundleId = (globalThis as any).Application.applicationId;
+    }
+
     // purposefully using SPLIT here and and not replace for CID to avoid cases where users don't know the schema
     // also only splitting on `/ipfs` to avoid cases where people pass non `/` terminated gateway urls
     return `${
       gateway.replace("{clientId}", clientId).split("/ipfs")[0]
-    }/ipfs/${cid}`;
+    }/ipfs/${cid}${bundleId ? `?bundleId=${bundleId}` : ""}`;
   }
   if (options.uri.startsWith("http")) {
     return options.uri;